+++ src/main/java/net/frozenblock/wilderwild/mixin/client/allay/AllayModelMixin.java
@@ -30,7 +30,6 @@ import net.minecraft.client.model.HierarchicalModel;
 import net.minecraft.world.entity.animal.allay.Allay;
 import org.objectweb.asm.Opcodes;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -39,22 +38,24 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(AllayModel.class)
 public abstract class AllayModelMixin extends HierarchicalModel<Allay> implements ArmedModel {
 
-	@Unique
-	private final AllayModel wilderWild$model = AllayModel.class.cast(this);
-
 	@ModifyExpressionValue(
 		method = "setupAnim(Lnet/minecraft/world/entity/animal/allay/Allay;FFFFF)V",
-		at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/animal/allay/Allay;isDancing()Z", ordinal = 0),
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/entity/animal/allay/Allay;isDancing()Z",
+			ordinal = 0
+		),
 		require = 0
 	)
 	private boolean wilderWild$alterDanceCheck(boolean original) {
-		return original && !EntityConfig.get().allay.keyframeAllayDance;
+		return original && !EntityConfig.Client.KEYFRAME_ALLAY_DANCE;
 	}
 
 	@Inject(
 		method = "setupAnim(Lnet/minecraft/world/entity/animal/allay/Allay;FFFFF)V",
 		at = @At(
-			value = "FIELD", target = "Lnet/minecraft/client/model/AllayModel;right_wing:Lnet/minecraft/client/model/geom/ModelPart;",
+			value = "FIELD",
+			target = "Lnet/minecraft/client/model/AllayModel;right_wing:Lnet/minecraft/client/model/geom/ModelPart;",
 			opcode = Opcodes.GETFIELD,
 			ordinal = 0,
 			shift = At.Shift.BEFORE
@@ -62,8 +63,8 @@ public abstract class AllayModelMixin extends HierarchicalModel<Allay> implement
 		require = 0
 	)
 	private void wilderWild$runKeyframeDance(Allay allay, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch, CallbackInfo info) {
-		if (EntityConfig.get().allay.keyframeAllayDance) {
-			this.wilderWild$model.animate(((WilderAllay) allay).wilderWild$getDancingAnimationState(), CustomAllayAnimations.DANCING, ageInTicks);
+		if (EntityConfig.Client.KEYFRAME_ALLAY_DANCE && allay instanceof WilderAllay wilderAllay) {
+			this.animate(wilderAllay.wilderWild$getDancingAnimationState(), CustomAllayAnimations.DANCING, ageInTicks);
 		}
 	}
 }