+++ src/main/java/net/frozenblock/wilderwild/mixin/client/mesoglea/BubbleColumnAmbientSoundHandlerMixin.java
@@ -35,18 +35,42 @@ import org.spongepowered.asm.mixin.injection.At;
 @Mixin(BubbleColumnAmbientSoundHandler.class)
 public class BubbleColumnAmbientSoundHandlerMixin {
 
-	@WrapOperation(method = "method_29714(Lnet/minecraft/world/level/block/state/BlockState;)Z", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/world/level/block/Block;)Z", ordinal = 0), require = 0)
+	@WrapOperation(
+		method = "method_29714(Lnet/minecraft/world/level/block/state/BlockState;)Z",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/world/level/block/Block;)Z",
+			ordinal = 0
+		),
+		require = 0
+	)
 	private static boolean wilderWild$filterWithMesoglea(BlockState state, Block block, Operation<Boolean> operation) {
 		return operation.call(state, block) || MesogleaBlock.hasBubbleColumn(state);
 	}
 
-	@WrapOperation(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/world/level/block/Block;)Z", ordinal = 0), require = 0)
-	private static boolean wilderWild$mesogleaAndBubbleCheck(BlockState state, Block block, Operation<Boolean> operation) {
+	@WrapOperation(
+		method = "tick",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/level/block/state/BlockState;is(Lnet/minecraft/world/level/block/Block;)Z",
+			ordinal = 0
+		),
+		require = 0
+	)
+	private boolean wilderWild$mesogleaAndBubbleCheck(BlockState state, Block block, Operation<Boolean> operation) {
 		return operation.call(state, block) || MesogleaBlock.hasBubbleColumn(state);
 	}
 
-	@WrapOperation(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/level/block/state/BlockState;getValue(Lnet/minecraft/world/level/block/state/properties/Property;)Ljava/lang/Comparable;", ordinal = 0), require = 0)
-	private static Comparable<Boolean> wilderWild$blockStateValueCheckWithMesoglea(BlockState state, Property<?> property, Operation<Comparable<Boolean>> operation) {
+	@WrapOperation(
+		method = "tick",
+		at = @At(
+			value = "INVOKE",
+			target = "Lnet/minecraft/world/level/block/state/BlockState;getValue(Lnet/minecraft/world/level/block/state/properties/Property;)Ljava/lang/Comparable;",
+			ordinal = 0
+		),
+		require = 0
+	)
+	private Comparable<Boolean> wilderWild$blockStateValueCheckWithMesoglea(BlockState state, Property<?> property, Operation<Comparable<Boolean>> operation) {
 		return state.hasProperty(BubbleColumnBlock.DRAG_DOWN) ? operation.call(state, property) : MesogleaBlock.isDraggingDown(state);
 	}