+++ src/main/java/net/frozenblock/wilderwild/mixin/client/wind/fallingleaves/FallingLeafParticleMixin.java
@@ -39,7 +39,15 @@ public class FallingLeafParticleMixin {
 		return AmbienceAndMiscConfig.CLOUD_MOVEMENT && ClientWindManager.shouldUseWind();
 	}
 
-	@ModifyExpressionValue(method = "tick", at = @At(value = "FIELD", target = "Lrandommcsomethin/fallingleaves/util/Wind;windX:F"), require = 0, remap = false)
+	@ModifyExpressionValue(
+		method = "tick",
+		at = @At(
+			value = "FIELD",
+			target = "Lrandommcsomethin/fallingleaves/util/Wind;windX:F"
+		),
+		require = 0,
+		remap = false
+	)
 	private float wilderWild$modifyWindX(float original) {
 		if (wilderWild$useWind()) {
 			return (float) ClientWindManager.windX * 0.7F;
@@ -47,7 +55,15 @@ public class FallingLeafParticleMixin {
 		return original;
 	}
 
-	@ModifyExpressionValue(method = "tick", at = @At(value = "FIELD", target = "Lrandommcsomethin/fallingleaves/util/Wind;windZ:F"), require = 0, remap = false)
+	@ModifyExpressionValue(
+		method = "tick",
+		at = @At(
+			value = "FIELD",
+			target = "Lrandommcsomethin/fallingleaves/util/Wind;windZ:F"
+		),
+		require = 0,
+		remap = false
+	)
 	private float wilderWild$modifyWindZ(float original) {
 		if (wilderWild$useWind()) {
 			return (float) ClientWindManager.windZ * 0.7F;