+++ CHANGELOG.md
@@ -4,4 +4,7 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 
 -----------------
 - Removed all but three of Wilder Wild's panoramas.
+- Cleaned up and optimized many mixins.
+- Significantly improved and optimized Warden swimming and its animations.
+- Optimized Allay animations.
 - Slightly optimized the mixin plugin.